+++ service.go
@@ -238,7 +238,7 @@ func (self *Service) GetDescription() (string, error) {
 	return LpstrToString(desc.Description), nil
 }
 
-func (self *Service) GetProcessID() (uint32, error) {
+func (self *Service) GetProcessID() (uint, error) {
 	var status wrappers.SERVICE_STATUS_PROCESS
 	size := uint32(unsafe.Sizeof(status))
 	err := wrappers.QueryServiceStatusEx(
@@ -250,7 +250,7 @@ func (self *Service) GetProcessID() (uint32, error) {
 	if err != nil {
 		return 0, NewWindowsError("QueryServiceStatusEx", err)
 	}
-	return status.ProcessId, nil
+	return uint(status.ProcessId), nil
 }
 
 func (self *Service) GetStatus() (*ServiceStatusInfo, error) {
